tweak is_writable check in db backup
git-svn-id: https://plugins.svn.wordpress.org/updraftplus/trunk@751786 b8457f37-d9ea-0310-8a92-e5e31aec5664

+++ methods/googledrive.php
@@ -188,13 +188,20 @@ class UpdraftPlus_BackupModule_googledrive {
 			$file_name = basename($file_path);
 			$updraftplus->log("$file_name: Attempting to upload to Google Drive");
 
+			$filesize = filesize($file_path);
+			$already_failed = false;
 			if ($available_quota != -1) {
-				$filesize = filesize($file_path);
 				if ($filesize > $available_quota) {
+					$already_failed = true;
 					$updraftplus->log("File upload expected to fail: file ($file_name) size is $filesize b, whereas available quota is only $available_quota b");
 					$updraftplus->log(sprintf(__("Account full: your %s account has only %d bytes left, but the file to be uploaded is %d bytes",'updraftplus'),__('Google Drive', 'updraftplus'), $available_quota, $filesize), 'error');
 				}
 			}
+			if (!$already_failed && $filesize > 10737418240) {
+				# 10Gb
+				$updraftplus->log("File upload expected to fail: file ($file_name) size is $filesize b (".round($filesize/1073741824, 4)." Gb), whereas Google Drive's limit is 10Gb (1073741824 bytes)");
+				$updraftplus->log(sprintf(__("Upload expected to fail: the %s limit for any single file is %s, whereas this file is %s Gb (%d bytes)",'updraftplus'),__('Google Drive', 'updraftplus'), '10Gb (1073741824)', round($filesize/1073741824, 4), $filesize), 'warning');
+			} 
 
 			$timer_start = microtime(true);
 			if ( $id = $this->upload_file( $file_path, $file_name, UpdraftPlus_Options::get_updraft_option('updraft_googledrive_remotepath')) ) {
